-（迁移）批量修复 26.3 渲染/粒子/界面/持久化 API

+++ src/main/java/alku/csrp/entity/CrudeParasiteEntity.java
@@ -16,7 +16,7 @@ public abstract class CrudeParasiteEntity extends PrimitiveParasiteEntity {
     }
 
     @Override
-    public boolean causeFallDamage(float distance, float damageMultiplier, DamageSource source) {
-        return distance >= 60.0F && super.causeFallDamage(distance, damageMultiplier, source);
+    public boolean causeFallDamage(double distance, float damageMultiplier, DamageSource source) {
+        return distance >= 60.0D && super.causeFallDamage(distance, damageMultiplier, source);
     }
 }
